Cleanup p2p code (#96)
