+++ main.go
@@ -198,7 +198,7 @@ func ontMain(ctx *cli.Context) {
 		os.Exit(1)
 	}
 	p2p.SetPID(p2pPID)
-	err = p2p.Start(true)
+	err = p2p.Start()
 	if err != nil {
 		log.Fatalf("p2p sevice start error %s", err)
 		os.Exit(1)
