+++ p2pserver/net/netserver/net_utils.go
@@ -1,0 +1,149 @@
+/*
+ * Copyright (C) 2018 The dad-go Authors
+ * This file is part of The dad-go library.
+ *
+ * The dad-go is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * The dad-go is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with The dad-go.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package netserver
+
+import (
+	"crypto/tls"
+	"crypto/x509"
+	"errors"
+	"io/ioutil"
+	"net"
+	"strconv"
+	"time"
+
+	"github.com/ontio/dad-go/common/config"
+	"github.com/ontio/dad-go/common/log"
+	"github.com/ontio/dad-go/p2pserver/common"
+)
+
+// createListener creates a net listener on the port
+func createListener(port uint16) (net.Listener, error) {
+	var listener net.Listener
+	var err error
+
+	isTls := config.Parameters.IsTLS
+	if isTls {
+		listener, err = initTlsListen(port)
+		if err != nil {
+			log.Error("initTlslisten failed")
+			return nil, errors.New("initTlslisten failed")
+		}
+	} else {
+		listener, err = initNonTlsListen(port)
+		if err != nil {
+			log.Error("initNonTlsListen failed")
+			return nil, errors.New("initNonTlsListen failed")
+		}
+	}
+	return listener, nil
+}
+
+//nonTLSDial return net.Conn with nonTls
+func nonTLSDial(addr string) (net.Conn, error) {
+	log.Debug()
+	conn, err := net.DialTimeout("tcp", addr, time.Second*common.DIAL_TIMEOUT)
+	if err != nil {
+		return nil, err
+	}
+	return conn, nil
+}
+
+//TLSDial return net.Conn with TLS
+func TLSDial(nodeAddr string) (net.Conn, error) {
+	CertPath := config.Parameters.CertPath
+	KeyPath := config.Parameters.KeyPath
+	CAPath := config.Parameters.CAPath
+
+	clientCertPool := x509.NewCertPool()
+
+	cacert, err := ioutil.ReadFile(CAPath)
+	cert, err := tls.LoadX509KeyPair(CertPath, KeyPath)
+	if err != nil {
+		return nil, err
+	}
+
+	ret := clientCertPool.AppendCertsFromPEM(cacert)
+	if !ret {
+		return nil, errors.New("failed to parse root certificate")
+	}
+
+	conf := &tls.Config{
+		RootCAs:      clientCertPool,
+		Certificates: []tls.Certificate{cert},
+	}
+
+	var dialer net.Dialer
+	dialer.Timeout = time.Second * common.DIAL_TIMEOUT
+	conn, err := tls.DialWithDialer(&dialer, "tcp", nodeAddr, conf)
+	if err != nil {
+		return nil, err
+	}
+	return conn, nil
+}
+
+//initNonTlsListen return net.Listener with nonTls mode
+func initNonTlsListen(port uint16) (net.Listener, error) {
+	log.Debug()
+	listener, err := net.Listen("tcp", ":"+strconv.Itoa(int(port)))
+	if err != nil {
+		log.Error("Error listening\n", err.Error())
+		return nil, err
+	}
+	return listener, nil
+}
+
+//initTlsListen return net.Listener with Tls mode
+func initTlsListen(port uint16) (net.Listener, error) {
+	CertPath := config.Parameters.CertPath
+	KeyPath := config.Parameters.KeyPath
+	CAPath := config.Parameters.CAPath
+
+	// load cert
+	cert, err := tls.LoadX509KeyPair(CertPath, KeyPath)
+	if err != nil {
+		log.Error("load keys fail", err)
+		return nil, err
+	}
+	// load root ca
+	caData, err := ioutil.ReadFile(CAPath)
+	if err != nil {
+		log.Error("read ca fail", err)
+		return nil, err
+	}
+	pool := x509.NewCertPool()
+	ret := pool.AppendCertsFromPEM(caData)
+	if !ret {
+		return nil, errors.New("failed to parse root certificate")
+	}
+
+	tlsConfig := &tls.Config{
+		Certificates: []tls.Certificate{cert},
+		RootCAs:      pool,
+		ClientAuth:   tls.RequireAndVerifyClientCert,
+		ClientCAs:    pool,
+	}
+
+	log.Info("TLS listen port is ", strconv.Itoa(int(port)))
+	listener, err := tls.Listen("tcp", ":"+strconv.Itoa(int(port)), tlsConfig)
+	if err != nil {
+		log.Error(err)
+		return nil, err
+	}
+	return listener, nil
+}
